Final 2023-10-23 commit

+++ src/Parsers/parseQuery.cpp
@@ -3,6 +3,7 @@
 #include <Interpreters/OpenTelemetrySpanLog.h>
 #include <Parsers/ParserQuery.h>
 #include <Parsers/ASTInsertQuery.h>
+#include <Parsers/ASTExplainQuery.h>
 #include <Parsers/Lexer.h>
 #include <Parsers/TokenIterator.h>
 #include <Common/StringUtils/StringUtils.h>
@@ -12,6 +13,8 @@
 #include <IO/WriteHelpers.h>
 #include <IO/WriteBufferFromString.h>
 #include <IO/Operators.h>
+
+#include <Parsers/HasQuery.h>
 
 
 namespace DB
@@ -235,8 +238,17 @@ ASTPtr tryParseQuery(
     size_t max_parser_depth)
 {
     const char * query_begin = _out_query_end;
+    // 改 
+    // if((_out_query_end[0]=='S'||_out_query_end[0]=='s')&&(_out_query_end[1]=='E'||_out_query_end[1]=='e')){
+    //     HasQuery::hasquery = true;
+    // }
+    // 改 09-01
+    // if((_out_query_end[0]=='S'||_out_query_end[0]=='s')&&(_out_query_end[1]=='T'||_out_query_end[1]=='t')){
+    //     HasQuery::hasquery = true;
+    // }
     Tokens tokens(query_begin, all_queries_end, max_query_size);
-    IParser::Pos token_iterator(tokens, max_parser_depth);
+    /// NOTE: consider use UInt32 for max_parser_depth setting.
+    IParser::Pos token_iterator(tokens, static_cast<uint32_t>(max_parser_depth));
 
     if (token_iterator->isEnd()
         || token_iterator->type == TokenType::Semicolon)
